+++ packages/ui/src/components/controller-export/index.ts
@@ -6,6 +6,6 @@
 // option/wire-protocol types — is INTERNAL to <ControllerToaster /> and is
 // intentionally NOT re-exported. The folder's *.stories.tsx import those
 // internals directly from their module files, not from this barrel.
-export { ControllerToaster } from "./controller-toaster";
-export type { ControllerNotificationTypes } from "./controller-toaster";
+export { ControllerToaster } from "./controller-toaster/controller-toaster";
+export type { ControllerNotificationTypes } from "./controller-toaster/controller-toaster";
 export type { ToastPosition } from "./types";
